update: (add next intl): set up file
