+++ message/en.json
@@ -1,0 +1,7 @@
+{
+    "Home": {
+      "HeroSection": {
+        "Greeting": "Hello Greeting!"
+      }
+    }
+  }
