+++ message/th.json
@@ -1,0 +1,7 @@
+{
+    "Home": {
+      "HeroSection": {
+        "Greeting": "สวัสดีครับ"
+      }
+    }
+  }
