+++ next.config.js
@@ -1,7 +1,21 @@
 /** @type {import('next').NextConfig} */
+const withNextIntl = require('next-intl/plugin')();
+
 const nextConfig = {
-    output: 'export',
-    images: { unoptimized: false }
+    async redirects() {
+        return [
+          {
+            source: '/old-blog/:slug',
+            destination: '/news/:slug', // Matched parameters can be used in the destination
+            permanent: true,
+          },
+        ]
+      },
+    output: 'standalone',
+    images: { unoptimized: false },
 }
 
-module.exports = nextConfig
+
+
+
+module.exports = withNextIntl(nextConfig);
