+++ package.json
@@ -28,6 +28,7 @@
     "gsap": "^3.12.2",
     "luxon": "^3.4.4",
     "next": "13.4.19",
+    "next-intl": "^3.4.1",
     "next-sitemap": "^4.2.3",
     "postcss": "8.4.30",
     "react": "18.2.0",
