+++ public/sitemap-0.xml
@@ -1,12 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://codework-tech.com/aboutus</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/privacy-policy</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/blog</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/team</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/contact-us</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/services</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/projects</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://codework-tech.com/blog/introduction-to-daisy-ui</loc><lastmod>2024-01-03T08:37:52.455Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/blog</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/aboutus</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/services</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/privacy-policy</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/en/contact-us</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/th/contact-us</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/projects</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/team</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/en</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/th</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/blog/article-1</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/blog/example</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/blog/introduction-to-daisy-ui</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://codework-tech.com/blog/understanding-how-ux-and-ui-work-together</loc><lastmod>2024-01-08T07:14:26.868Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
