+++ src/app/[locale]/contact-us/page.tsx
@@ -5,9 +5,12 @@ import { GTM_PRODUCTION, LOCAL_STORAGE_PDPA_KEY } from "@/config/environment";
 import FormPage from "@/features/ContactPage/FormPage";
 import PDPAPopup from "@/features/PAPAPopup/PDPAPopup";
 import { Metadata } from "next";
+
 import { Poppins } from "next/font/google";
 import { useEffect, useState } from "react";
 import TagManager from "react-gtm-module";
+import {useTranslations} from 'next-intl';
+import { unstable_setRequestLocale } from 'next-intl/server';
 
 const poppinsFont = Poppins({
   weight: ["100", "300", "500", "700"],
@@ -44,12 +47,22 @@ export const metadata: Metadata = {
   },
 };
 
-export default function ContactUsPage() {
+export default function ContactUsPage({params: {locale}}:any) {
+  unstable_setRequestLocale(locale);
 
+  const t = useTranslations('Home');
   return (
     <main className={` bg-neutral-1000 ${poppinsFont.className}`}>
+      <h1 className="text-white">{t('HeroSection.Greeting')}</h1>
       <FormPage />
       <SocialContactFloating />
     </main>
   );
 }
+
+const locales = ['en', 'th'];
+export function generateStaticParams() {
+  return locales.map((locale) => ({locale}));
+} 
+
+
